requirements.txt
torch
torchvision
fastapi
pillow
streamlit
requests
pandas
plotly
matplotlib
seaborn
numpy
scikit-learn
tqdm